+++ src/data/landing.js
@@ -1,0 +1,13 @@
+const landing = {
+  name: 'Artem Moroz',
+  src: 'https://avatars.githubusercontent.com/u/88348091?v=4',
+  introduction: 'React Developer',
+  greeting:
+    'Hello! I am Artem Moroz, a frontend developer specializing in React.',
+  bio: `Welcome to my portfolio!
+      Here you will find my projects, expertise, and articles related to React development.
+      Join me to learn new things, share experiences, and ask questions.
+      Welcome to the world of React!`,
+};
+
+export default landing;
