+++ src/data/links.js
@@ -1,0 +1,19 @@
+const links = [
+  {
+    name: 'Projects',
+    id: 'projects',
+    url: '#projects',
+  },
+  {
+    name: 'Certificates',
+    id: 'certificates',
+    url: '#certificates',
+  },
+  {
+    name: 'Contact Me',
+    id: 'contact',
+    url: '#contact',
+  },
+];
+
+export default links;
