+++ src/data/projects.js
@@ -1,0 +1,21 @@
+import photo1 from '../assets/little_lemon_1.png';
+import photo2 from '../assets/little_lemon_1.png';
+
+const projects = [
+  {
+    title: 'Little Lemon',
+    description:
+      'The project includes a landing page and reservation form for a restaurant, providing essential information and enabling easy table bookings.',
+    getImageSrc: () => photo1,
+    url: 'https://github.com/moroz-art-dev/little-lemon',
+  },
+  {
+    title: 'Life search',
+    description:
+      'A scrollable bottom sheet with virtualisation support, native animations at 60 FPS and fully implemented in JS land 🔥️',
+    getImageSrc: () => photo2,
+    url: 'https://github.com/moroz-art-dev/little-lemon',
+  },
+];
+
+export default projects;
